Refactor sorting for clarity
